Vicat/glass temp from -100 to 500+
Some materials do not have a glass transition temperature, they have a
vicat softening point. I think it's either/or, so we can probably use
one field for both, though we may need an extra bit to say which one it
is.

The range for this supposedly needs to start below zero. Not sure yet
whether we need to specify exact values below zero. If not, we can shave
off one bit and make the range start at -1.